+++ SistemaCajaRegistradora/Controllers/UsuarioController.cs
@@ -20,34 +20,24 @@ namespace SistemaCajaRegistradora.Controllers
 
         [HttpGet]
         [ActionName("Listar")]
-        [Autorizacion(idoperacion:12)]
-        public ActionResult Listar()
-        {
-            return View();
-        }
+        [Autorizacion(idoperacion: 12)]
+        public ActionResult Listar() => View();
 
         [HttpGet]
         [ActionName("getUsuarios")]
         public JsonResult getUsuarios()
         {
             db.Configuration.LazyLoadingEnabled = false;
-            List<vmUsuario> usuariosList = new List<vmUsuario>();
-            vmUsuario vmUsuario;
-            var usuarios = db.Usuarios.Include(u => u.Imagen).Include(u => u.Role).ToArray();
-            foreach (var usuario in usuarios)
+            List<vmUsuario> usuariosList = db.Usuarios.Include(u => u.Imagen).Include(u => u.Role).Select(u => new vmUsuario()
             {
-                vmUsuario = new vmUsuario()
-                {
-                    id = usuario.id,
-                    nombreCajero = usuario.nombre.Trim() + " " + usuario.apellido.Trim(),
-                    nombreUsuario = usuario.nombreUsuario.Trim(),
-                    rutaImg = usuario.Imagen.ruta,
-                    rol = usuario.Role.rol.Trim(),
-                    stateConexion = usuario.conectado,
-                    solrespass = usuario.solrespass
-                };
-                usuariosList.Add(vmUsuario);
-            }
+                id = u.id,
+                nombreCajero = u.nombre.Trim() + " " + u.apellido.Trim(),
+                nombreUsuario = u.nombreUsuario.Trim(),
+                rutaImg = u.Imagen.ruta.Trim(),
+                rol = u.Role.rol.Trim(),
+                stateConexion = u.conectado,
+                solrespass = u.solrespass
+            }).ToList();
 
             return Json(new
             {
@@ -58,31 +48,27 @@ namespace SistemaCajaRegistradora.Controllers
         [HttpGet]
         [ActionName("AgregarForms")]
         [Autorizacion(idoperacion: 13)]
-        public PartialViewResult AgregarForms()
-        {
-            return PartialView("_formsUsuario");
-        }
+        public PartialViewResult AgregarForms() => PartialView("_formsUsuario");
 
         [HttpPost]
         [ActionName("AgregarUsuario")]
         [Autorizacion(idoperacion: 13)]
         public JsonResult AgregarUsuario(Usuario usuario)
         {
-            if (usuario.rolid != 1)
+            if (usuario.rolid == 1)
             {
-                usuario.rolid = 2; //Asignamos el rol de cajero
-                usuario.clave = Encrypt.GetSHA256(usuario.clave);
-                usuario.imagenid = 2; //Se define una imagen por defecto
-                usuario.fecha_creacion = DateTime.UtcNow;
-                usuario.fecha_modificacion = usuario.fecha_creacion;
-                usuario.conectado = false;
-                db.Usuarios.Add(usuario);
-                int n = db.SaveChanges();
-                return Json(n, JsonRequestBehavior.AllowGet);
+                //Si existe un administrador, enviar -1 que significa error de creacion de usuario
+                return Json(-1, JsonRequestBehavior.AllowGet);
             }
-            //Si existe un administrador, enviar -1 que significa error de creacion de usuario
-            return Json(-1, JsonRequestBehavior.AllowGet);
-            
+            usuario.rolid = 2; //Asignamos el rol de cajero
+            usuario.clave = Encrypt.GetSHA256(usuario.clave);
+            usuario.imagenid = 2; //Se define una imagen por defecto
+            usuario.fecha_creacion = DateTime.UtcNow;
+            usuario.fecha_modificacion = usuario.fecha_creacion;
+            usuario.conectado = false;
+            db.Usuarios.Add(usuario);
+            int n = db.SaveChanges();
+            return Json(n, JsonRequestBehavior.AllowGet);
         }
 
         [HttpGet]
@@ -90,11 +76,18 @@ namespace SistemaCajaRegistradora.Controllers
         [Autorizacion(idoperacion: 15)]
         public PartialViewResult formsEditar(int? id)
         {
-            var usuario = db.Usuarios.Include(u => u.Role)
-                            .Where(u => u.id == id).FirstOrDefault();
-            var roles = db.Roles.ToList();
-            quitarEspaciosVacios(usuario);
+            var user_rol_data = db.Usuarios.Include(u => u.Role).ToArray();
+
+            var usuario = user_rol_data.FirstOrDefault(u => u.id == id);
+            var roles = user_rol_data.Select(r => new Role()
+            {
+                id = r.Role.id,
+                rol = r.Role.rol,
+                rol_operacion = r.Role.rol_operacion
+            }).ToList();
             ViewBag.rolesId = new SelectList(roles, "id", "rol", usuario.rolid);
+
+            quitarEspaciosVacios(usuario);
             return PartialView("_formsUsuario", usuario);
         }
 
@@ -108,7 +101,7 @@ namespace SistemaCajaRegistradora.Controllers
             {
                 usuario.conectado = true;
             }
-            usuario.fecha_modificacion = DateTime.UtcNow;
+            usuario.fecha_modificacion = DateTime.Now;
             db.Entry(usuario).State = EntityState.Modified;
             int n = db.SaveChanges();
             return Json(n, JsonRequestBehavior.AllowGet);
@@ -119,7 +112,9 @@ namespace SistemaCajaRegistradora.Controllers
         [Autorizacion(idoperacion:16)]
         public PartialViewResult formsEliminar(int? id)
         {
-            var usuario = db.Usuarios.Include(u=>u.Role).Where(u=>u.id == id).FirstOrDefault();
+            var usuario = db.Usuarios.Include(u=>u.Role).FirstOrDefault(u => u.id == id);
+            if (usuario == null) return PartialView(null);
+
             quitarEspaciosVacios(usuario);
             return PartialView("_formsEliminar", usuario);
         }
@@ -134,16 +129,33 @@ namespace SistemaCajaRegistradora.Controllers
 
             try
             {
-                var usuario = db.Usuarios.Include(u => u.Imagen)
-                .Where(u => u.id == id).FirstOrDefault();
+                var usuario_imagen_data = db.Usuarios.Include(u => u.Imagen).ToArray();
+                var imagenes = usuario_imagen_data.Select(i => new Imagen()
+                {
+                    id = i.Imagen.id,
+                    nombre = i.Imagen.nombre.Trim(),
+                    ruta = i.Imagen.ruta.Trim(),
+                }).ToArray();
+
+                var usuario = usuario_imagen_data.FirstOrDefault(u => u.id == id);
+
+                if (usuario == null)
+                {
+                    return Json(new
+                    {
+                        n,
+                        nameFile,
+                        idimg = 0
+                    }, JsonRequestBehavior.AllowGet);
+                }
+                db.Usuarios.Remove(usuario);
+
                 nameFile = usuario.Imagen.nombre.Trim();
                 long idimg = usuario.imagenid;
-                db.Imagens.Find(idimg);
-                db.Usuarios.Remove(usuario);
-                if (idimg != 2)
-                {
-                    db.Usuarios.Remove(usuario);
-                }
+                var img = imagenes.FirstOrDefault(i => i.id == idimg);
+                if (img != null && idimg != 2) 
+                    db.Imagens.Remove(img);
+
                 n = db.SaveChanges();
                 return Json(new
                 {
@@ -168,15 +180,10 @@ namespace SistemaCajaRegistradora.Controllers
         [ActionName("formsCambiarPass")]
         public PartialViewResult formsCambiarPass()
         {
-            if (Session["User"] != null)
-            {
-                Usuario user = (Usuario)Session["User"];
-                return PartialView("_formsCambiarPass", user);
-            }
-            else
-            {
-                return PartialView("_formsRestablecerPass");
-            }
+            if (Session["User"] == null) return PartialView("_formsRestablecerPass");
+
+            Usuario user = (Usuario)Session["User"];
+            return PartialView("_formsCambiarPass", user);
         }
 
         [HttpPost]
@@ -200,7 +207,9 @@ namespace SistemaCajaRegistradora.Controllers
         public PartialViewResult formsImagen(int? id)
         {
             Session["idUser"] = id;
-            var usuario = db.Usuarios.Where(u => u.id == id).FirstOrDefault();
+            var usuario = db.Usuarios.FirstOrDefault(u => u.id == id);
+            if (usuario == null) return PartialView(null);
+
             return PartialView("_formsImagenUsuario", usuario);
         }
 
@@ -211,75 +220,72 @@ namespace SistemaCajaRegistradora.Controllers
         {
             using (TransactionScope scope = new TransactionScope())
             {
-                using (ModelData db1 = new ModelData())
+                string msgSuccess = "Imagen subida correctamente";
+                string msgError = "Error de subida de archivo";
+                try
                 {
-                    string msgSuccess = "Imagen subida correctamente";
-                    string msgError = "Error de subida de archivo";
-                    try
+                    if (downloadURL.Equals(string.Empty) && nameFile.Equals(string.Empty))
+                        throw new Exception(msgError);
+
+                    Imagen img = new Imagen()
                     {
-                        if (downloadURL.Equals(string.Empty) && nameFile.Equals(string.Empty))
-                            throw new Exception(msgError);
+                        nombre = nameFile,
+                        ruta = downloadURL,
+                    };
+                    db.Imagens.Add(img);
+                    int n = db.SaveChanges();
 
-                        Imagen img = new Imagen()
-                        {
-                            nombre = nameFile,
-                            ruta = downloadURL,
-                        };
-                        db1.Imagens.Add(img);
-                        int n = db1.SaveChanges();
+                    if (n == 0)
+                        throw new Exception(msgError);
 
-                        if (n == 0)
-                            throw new Exception(msgError);
+                    int idUsuario = (int)Session["idUser"];
 
-                        int idUsuario = (int)Session["idUser"];
-
-                        //Cambiar la imagen a la variable de session
-                        var user = (Usuario)Session["User"];
-                        if (user.id == idUsuario)
-                        {
-                            user.imagenid = img.id;
-                            Session["User"] = user;
-                        }
-
-                        var usuario = db1.Usuarios.Find(idUsuario);
-                        string nombreImg = usuario.Imagen.nombre;
-                        long idimg = usuario.imagenid;
-
-                        usuario.imagenid = img.id;
-                        usuario.fecha_modificacion = DateTime.Now;
-                        db1.Entry(usuario).State = EntityState.Modified;
-                        n = db1.SaveChanges();
-                        if (n == 0)
-                            throw new Exception(msgError);
-
-                        //Borrar registro de la otra imagen
-                        if (idimg != 2)
-                        {
-                            var imagen = db1.Imagens.Find(idimg);
-                            db1.Imagens.Remove(imagen);
-                            db1.SaveChanges();
-                        }
-
-                        scope.Complete();
-
-                        return Json(new
-                        {
-                            status = "success",
-                            msg = msgSuccess,
-                            idimg,
-                            nombreImg
-                        },JsonRequestBehavior.AllowGet);
-                    }
-                    catch (Exception)
+                    //Cambiar la imagen a la variable de session
+                    var user = (Usuario)Session["User"];
+                    if (user.id == idUsuario)
                     {
-                        return Json(new
-                        {
-                            status = "error",
-                            msg = msgError,
-                            idimg = 0,
-                            nombreImg = ""
-                        },JsonRequestBehavior.AllowGet);
+                        user.imagenid = img.id;
+                        Session["User"] = user;
                     }
+
+                    var usuario = db.Usuarios.Find(idUsuario);
+                    string nombreImg = usuario.Imagen.nombre;
+                    long idimg = usuario.imagenid;
+
+                    usuario.imagenid = img.id;
+                    usuario.fecha_modificacion = DateTime.Now;
+                    db.Entry(usuario).State = EntityState.Modified;
+                    n = db.SaveChanges();
+                    if (n == 0)
+                        throw new Exception(msgError);
+
+                    //Borrar registro de la otra imagen
+                    if (idimg != 2)
+                    {
+                        var imagen = db.Imagens.Find(idimg);
+                        db.Imagens.Remove(imagen);
+                        db.SaveChanges();
+                    }
+
+                    scope.Complete();
+
+                    return Json(new
+                    {
+                        status = "success",
+                        msg = msgSuccess,
+                        idimg,
+                        nombreImg
+                    }, JsonRequestBehavior.AllowGet);
+                }
+                catch (Exception)
+                {
+                    return Json(new
+                    {
+                        status = "error",
+                        msg = msgError,
+                        idimg = 0,
+                        nombreImg = ""
+                    }, JsonRequestBehavior.AllowGet);
                 }
             }
         }
@@ -318,14 +324,17 @@ namespace SistemaCajaRegistradora.Controllers
         [Autorizacion(idoperacion: 20)]
         public ActionResult RestablecerPass(string user, string clave)
         {
-            int n = 0;
-            var usuario = db.Usuarios.Where(u => u.nombreUsuario == user).FirstOrDefault();
+            var usuario = db.Usuarios.FirstOrDefault(u => u.nombreUsuario == user);
+            if (usuario == null)
+            {
+                return View();
+            }
             usuario.solrespass = false;
             usuario.clave = clave;
             usuario.clave = Encrypt.GetSHA256(usuario.clave);
             usuario.fecha_modificacion = DateTime.Now;
             db.Entry(usuario).State = EntityState.Modified;
-            n = db.SaveChanges();
+            db.SaveChanges();
             return RedirectToAction("Listar");
         }
 
@@ -335,13 +344,11 @@ namespace SistemaCajaRegistradora.Controllers
         {
             if (id != null)
             {
-                var usuario = db.Usuarios.Where(u => u.id == id).FirstOrDefault();
-                if (usuario != null)
-                {
-                    ViewBag.code = Encrypt.GetSHA256(usuario.fecha_creacion.Value.Ticks.ToString());
-                    return PartialView("_formObtenerBarcode", usuario);
-                }
-                return PartialView(null);
+                var usuario = db.Usuarios.FirstOrDefault(u => u.id == id);
+                if (usuario == null) return PartialView(null);
+
+                ViewBag.code = Encrypt.GetSHA256(usuario.fecha_creacion.Value.Ticks.ToString());
+                return PartialView("_formObtenerBarcode", usuario);
             }
             return PartialView(null);
         }
@@ -359,14 +366,7 @@ namespace SistemaCajaRegistradora.Controllers
 
             if (usuario.id == -1)
             {
-                if (user == null)
-                {
-                    valid = true;
-                }
-                else
-                {
-                    valid = false;
-                }
+                valid = user == null ? true : false;
 
                 return Json(new
                 {
@@ -374,28 +374,22 @@ namespace SistemaCajaRegistradora.Controllers
                     createNameUser = usuario.nombreUsuario,
                 }, JsonRequestBehavior.AllowGet);
             }
-            else
+            while (!valid)
             {
-                while (!valid)
+                valid = user == null ? true : false;
+
+                if (!valid)
                 {
-                    if (user == null)
-                    {
-                        valid = true;
-                    }
-                    else
-                    {
-                        int numero = new Random().Next(10, 100);
-                        usuario.nombreUsuario += numero;
-                        user = db.Usuarios.Where(u => u.nombreUsuario == usuario.nombreUsuario).FirstOrDefault();
-                        valid = false;
-                    }
+                    int numero = new Random().Next(10, 100);
+                    usuario.nombreUsuario += numero;
+                    user = db.Usuarios.FirstOrDefault(u => u.nombreUsuario == usuario.nombreUsuario);
                 }
-                return Json(new
-                {
-                    nombreUsuario = "",
-                    createNameUser = usuario.nombreUsuario,
-                }, JsonRequestBehavior.AllowGet);
             }
+            return Json(new
+            {
+                nombreUsuario = "",
+                createNameUser = usuario.nombreUsuario,
+            }, JsonRequestBehavior.AllowGet);
         }
         [HttpPost]
         [ActionName("verificarClave")]
@@ -404,14 +398,9 @@ namespace SistemaCajaRegistradora.Controllers
             usuario.clave = Encrypt.GetSHA256(usuario.clave);
             var user = (Usuario)Session["User"];
             user.clave = user.clave.Trim();
-            if (user.clave==usuario.clave)
-            {
-                return Json(new { success = true },JsonRequestBehavior.AllowGet);
-            }
-            else
-            {
-                return Json(new { success = false }, JsonRequestBehavior.AllowGet);
-            }
+            return user.clave == usuario.clave ? 
+                Json(new { success = true }, JsonRequestBehavior.AllowGet) :
+                Json(new { success = false }, JsonRequestBehavior.AllowGet);
         }
 
         [HttpGet]
@@ -419,18 +408,13 @@ namespace SistemaCajaRegistradora.Controllers
         public JsonResult getSesion()
         {
             var user = (Usuario)Session["User"];
-            if (user != null)
+            if (user == null) return Json(null, JsonRequestBehavior.AllowGet);
+
+            return Json(new
             {
-                return Json(new
-                {
-                    nombreuser = user.nombre.Trim() + ' ' + user.apellido.Trim() + " (" + user.nombreUsuario.Trim() + ")",
-                    imgruta = user.Imagen.ruta.Trim()
-                }, JsonRequestBehavior.AllowGet);
-            }
-            else
-            {
-                return Json(null, JsonRequestBehavior.AllowGet);
-            }
+                nombreuser = user.nombre.Trim() + ' ' + user.apellido.Trim() + " (" + user.nombreUsuario.Trim() + ")",
+                imgruta = user.Imagen.ruta.Trim()
+            }, JsonRequestBehavior.AllowGet);
         }
         private void quitarEspaciosVacios(Usuario usuario)
         {
